Fixed bug 1352 - timidity MIDI playback plays drums that should be musical notes
manuel.montezelo@gmail.com 2011-12-28 11:42:44 PST
Hi,

I am forwarding the bug reported in Debian:
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=575575

The submitter requests the following (patch attached here as well):
"When using timidity MIDI playback, music played in MIDI channels 10 and
16 are interpreted as drums.  Really, only channel 10 should default to
drums, according to the MIDI standard; drums on channel 16 are a
Microsoft substandard. :)

To fix this misfeature, please apply the attached patch.  Contrary to
what the comment above the affected line says, this is *not* a runtime
option in SDL_mixer, as far as I can tell."